Initial work for wiring in the `acars_vdlm2_parser` library.
Made the internal messaging pass along `String` for everything. Will have to see how it goes.

Fixed a unit test as well.

+++ src/generics.rs
@@ -63,7 +63,7 @@ pub struct OutputServerConfig {
     pub listen_tcp: Option<Vec<u16>>,
     pub receive_tcp: Option<Vec<String>>,
     pub receive_zmq: Option<Vec<String>>,
-    pub reassembly_window: u64,
+    pub reassembly_window: f64,
 }
 
 impl OutputServerConfig {
@@ -72,7 +72,7 @@ impl OutputServerConfig {
         listen_tcp: &Option<Vec<u16>>,
         receive_tcp: &Option<Vec<String>>,
         receive_zmq: &Option<Vec<String>>,
-        reassembly_window: &u64,
+        reassembly_window: &f64,
     ) -> Self {
         Self {
             listen_udp: listen_udp.clone(),
